Phase 6: Split test_dataset.py into 10 focused modules
Controlled reorganization preserving all 83 tests:
- test_dataset_creation.py   (9 tests)
- test_dataset_slicing.py    (14 tests)
- test_dataset_masking.py    (8 tests)
- test_dataset_units.py      (3 tests)
- test_dataset_coords.py     (8 tests)
- test_dataset_metadata.py   (6 tests)
- test_dataset_complex.py    (5 tests)
- test_dataset_math.py       (8 tests)
- test_dataset_squeeze.py    (4 tests)
- test_dataset_regressions.py (5 tests)

No test logic changed. Before/after: 83 tests, 83 passed, runtime ~50s.
